move swift tests to submodule
