+++ codegen-v2/src/tests/swift/samples/class.input.yaml
