+++ codegen-v2/src/tests/swift/samples/class.output.swift
