+++ codegen-v2/src/tests/swift/samples/enum.input.yaml
