+++ codegen-v2/src/tests/swift/samples/enum_extension.output.swift
