+++ codegen-v2/src/tests/swift/samples/enum_private.input.yaml
