+++ codegen-v2/src/tests/swift/samples/non-associated.input.yaml
