+++ codegen-v2/src/tests/swift/samples/non-associated.output.swift
