+++ codegen-v2/src/tests/swift/samples/optional.input.yaml
