+++ codegen-v2/src/tests/swift/samples/private_class.output.swift
