+++ codegen-v2/src/tests/swift/samples/struct.input.yaml
