+++ codegen-v2/src/tests/swift/samples/struct.output.swift
